Aligned equation numbers (#28)
* Align equation numbers.
Make the equation numbers of environments `equation` and `eqnarray`
align with the right (default) or left (document option "leqno")
margins, thus duplicating standard LaTeX behavior.

* Fix skrewed vertical alignment of equation numbers in `eqnarray`.

When the equation numbers in environment `eqnarray` are floated to the
right or to the left they are also shifted down by half of a quad.
Why?  This patch undoes the shift and realigns the equation numbers
with their associated equations.