New generator for horizontal space (HTML only)
Move code of do_space from "latexscan.mll" to all modules that use it
(e.g. Text) with the exception of Html.  There implement a fancy
algorithm to translate `\hspace` and `\hspace*` to series of Unicode
spaces.

The new function for HTML, HorizontalSpace.approximate_hspace closely
approximates the `length` argument with a minimum number of space
characters (of different widths).  As Length.t defines `Pixel of int`
rounding errors are unavoidable and add up to as much as 1px.

The difference between `\kern` and `\hskip` is modeled by careful
injection of zero-width spaces and zero-width joiners.