Rectify some horizontal space issues (#35)
* Introduce predefined horizontal spaces
with their most closely matching Unicode counterparts.  Care was taken
to distinguish `\enspace' (which ties) and `\enskip' (which does not
tie).

* Prepare for rewrite of horizontal-space generator.

Rename some identifiers to match their actual role.  The most visible
changes are:
    Text.put_nbsp  ->  Text.put_hspace
    Length.No  ->  Length.NotALength
    Length.font  ->  Length.base_font_size

Improvements and functional changes in module "Length":
  * Add some less common dimension units understood by TeX ("bp",
    "cc", "dd", "sp").
  * Split conversion of `em' and `ex'.  Now 1ex equals 0.5em.
  * Introduce the fixed conversion ratio from printer points (pt) to
    pixels (px) according to W3: 1px is equal to 0.75pt.  See
        https://www.w3.org/TR/CSS22/syndata.html#length-units
    and also
        https://drafts.csswg.org/css-values-3/#absolute-lengths
  * Change the base_font_size from 10 to 12 following the statements
    in
        https://developer.mozilla.org/en-US/docs/Web/CSS/font-size
    in subsection Description/ems.

* New generator for horizontal space (HTML only)

Move code of do_space from "latexscan.mll" to all modules that use it
(e.g. Text) with the exception of Html.  There implement a fancy
algorithm to translate `\hspace` and `\hspace*` to series of Unicode
spaces.

The new function for HTML, HorizontalSpace.approximate_hspace closely
approximates the `length` argument with a minimum number of space
characters (of different widths).  As Length.t defines `Pixel of int`
rounding errors are unavoidable and add up to as much as 1px.

The difference between `\kern` and `\hskip` is modeled by careful
injection of zero-width spaces and zero-width joiners.

* Cleanup code and make it usable in the real world.

Move the whole "zoo" of unicode spaces needed by approximate_hspace to
module OutUnicode.

Downgrade exception for negative length argument to approximate_hspace
to a warning.

* Allow for non-literal aguments of `\hspace' and `\hspace*'.