Fix "multiple definition" bug in \@printindex and \printindex. (#42)
Macros \@printindex and \printindex of package imakeidx.hva define the
label `section@the@hevea@index'.  Multiple indices generate multiple
definitions for this label.  The result is an invalid HTML file.

This patch fixes the problem by appending the index name to the label,
which results in `section@the@hevea@index@default' for the default
case.