Emulate "\discretionary" and implement "\-". (#22)
* Emulate \discretionary and implement \-.

Some browser interpret the zero-width space character, some the HTML5
tag `wbr` and presumable others do not understand either.  So, we have
introduced an extra level of indirection called
`\word@break@opportunity`, which can be redefined by the user to get
the desired translation of `\discretionary{}{}{}`.

TeX macro `\-` directly maps to the soft hyphen.

* Added the breakable hyphen as translation of \discretionary{-}{}{-}.
Replaced \ifx-based implementation with \ifthenelse-based.

* Add symbolic name for `word joiner` character.
It comes in handy when we define a ligature-breaking macro, because
`\soft@hyphen` alone does not break a ligature, but
`\word@joiner\soft@hyphen` does (and enables hyphenation at the
syllable boundary as desired).

* Revert "Add symbolic name for `word joiner` character."
The character is just wrong.  We need zwnj, the zero-width non joiner.

This reverts commit 24408d1e7ff380d57fd02721d5bd650d9fae8cdf.

* Introduce macro `\zero@width@non@joiner' for &zwnj;
This definition will be useful when we build ligature-breaking macros.